Refactored controller. Fixed database cleaner development issue (incorrect RACK_ENV value)
