+++ app/data_mapper_setup.rb
@@ -1,13 +1,8 @@
-require 'data_mapper'
-require 'dm-postgres-adapter'
+require "data_mapper"
+require "dm-postgres-adapter"
 
-require_relative 'models/user'
+require_relative "models/user"
 
-# set up a connection with the database
-DataMapper.setup(:default, ENV['DATABASE_URL'] || "postgres://localhost/chitter_#{ENV["RACK_ENV"]}")
-
-# checking that everything we wrote / the way we structured it is correct
+DataMapper.setup(:default, ENV["DATABASE_URL"] || "postgres://localhost/chitter_#{ ENV["RACK_ENV"] }")
 DataMapper.finalize
-
-# build the columsn and rows
 DataMapper.auto_upgrade!
