+++ spec/spec_helper.rb
@@ -1,12 +1,13 @@
-ENV['RACK_ENV'] = 'test'
-require File.join(File.dirname(__FILE__), '..', 'app/app.rb')
+ENV["RACK_ENV"] = "test"
 
-require 'capybara'
-require 'capybara/rspec'
-require 'rspec'
-require 'coveralls'
-require 'simplecov'
-require 'database_cleaner'
+require File.join(File.dirname(__FILE__), "..", "app/app.rb")
+
+require "capybara"
+require "capybara/rspec"
+require "rspec"
+require "coveralls"
+require "simplecov"
+require "database_cleaner"
 
 SimpleCov.formatters = [
   SimpleCov::Formatter::HTMLFormatter,
@@ -27,15 +28,19 @@ RSpec.configure do |config|
     mocks.verify_partial_doubles = true
   end
 
+  RSpec.configure do |config|
+  # Everything in this block runs once before all the tests run
   config.before(:suite) do
     DatabaseCleaner.strategy = :transaction
     DatabaseCleaner.clean_with(:truncation)
   end
 
+  # Everything in this block runs once before each individual test
   config.before(:each) do
     DatabaseCleaner.start
   end
 
+  # Everything in this block runs once after each individual test
   config.after(:each) do
     DatabaseCleaner.clean
   end
